Added user_online policy

+++ app/controllers/user_onlines_controller.rb
@@ -2,6 +2,7 @@ class UserOnlinesController < ApplicationController
   def create
     @user = current_user
     @useronline = UserOnline.new(user:  @user, status: "Available", active: true)
+    authorize @useronline
     if @useronline.save
       redirect_to root_path
     else
@@ -13,6 +14,7 @@ class UserOnlinesController < ApplicationController
   def destroy
     @user = find_user
     @online = UserOnline.find_by(user: @user)
+    authorize @online
     if @online.destroy
       redirect_to root_path
     else
